Revert default precision to float32 for all runs

+++ src/orchestr_ai/postprocessing/calculators/mace_calculator.py
@@ -331,7 +331,7 @@ class ReconstructedMACECalculator(MACECalculator):
         kwargs["head"] = self.base_head
 
         # Initialize base MACECalculator using 'models' parameter
-        dtype = kwargs.pop("default_dtype", "float64")
+        dtype = kwargs.pop("default_dtype", "float32")
         super().__init__(models=[model], device=str(device), default_dtype=dtype, **kwargs)
 
         # Set default head to base_head initially
